Replaced all floats with doubles
-All instances of float replaced with double
-All instances of Vector[0-9]+f were replaced with Vector[0-9]+d
-Anything else should've also been replaced

+++ src/Body.cpp
@@ -1,33 +1,33 @@
 #include "Body.h"
 
-void Body::setMass(float newMass){
+void Body::setMass(double newMass){
 	mass = std::abs(newMass);
 }
-void Body::setRadius(float newRadius){
+void Body::setRadius(double newRadius){
 	radius = std::abs(newRadius);
 }
 
 
-Body::Body(Eigen::Vector3f posInit, Eigen::Vector3f velInit, float massInit, float radiusInit){
+Body::Body(Eigen::Vector3d posInit, Eigen::Vector3d velInit, double massInit, double radiusInit){
 	position = posInit;
 	velocity = velInit;
 	mass = std::abs(massInit);
 	radius = std::abs(radiusInit);
 }
 
-Eigen::Vector3f Body::getPos(){
+Eigen::Vector3d Body::getPos(){
 	return position;
 }
-Eigen::Vector3f Body::getVel(){
+Eigen::Vector3d Body::getVel(){
 	return velocity;
 }
-Eigen::Vector3f Body::getAcc(){
+Eigen::Vector3d Body::getAcc(){
 	return acceleration;
 }
-float Body::getMass(){
+double Body::getMass(){
 	return mass;
 }
-float Body::getRadius(){
+double Body::getRadius(){
 	return radius;
 }
 
@@ -51,12 +51,12 @@ void Body::removeThruster(int thrusterNo){
 	Thrusters.erase(Thrusters.begin()+(thrusterNo-1));
 }
 
-float Body::getSensorValue(int sensorNo){
+double Body::getSensorValue(int sensorNo){
 	return Sensors[sensorNo-1]->getReading();
 }
 
-Eigen::Vector3f Body::calculateThrustVector(){
-	Eigen::Vector3f plc(0,0,0);
+Eigen::Vector3d Body::calculateThrustVector(){
+	Eigen::Vector3d plc(0,0,0);
 	//for(Thruster thrust:Thrusters){plc+=thrust.getThrust();} //I can dream, Harold
 	for(std::vector<Thruster>::iterator itr = Thrusters.begin(); itr!=Thrusters.end(); ++itr){
 		plc+=itr->viewThrust();
@@ -64,7 +64,7 @@ Eigen::Vector3f Body::calculateThrustVector(){
 	return plc;
 }
 
-void Body::Update(float dt){
+void Body::Update(double dt){
 	acceleration = this->calculateThrustVector()/this->getMass();
 	position+=dt*velocity+dt*dt/2*acceleration;
 	velocity+=dt*acceleration;
