+++ src/Sensor.h
@@ -9,18 +9,18 @@ class Body;
 class Sensor{
 public:
 	virtual void updateValue()=0;
-	float getValue();
-	float getReading();
+	double getValue();
+	double getReading();
 
-	void setVariance(float newVariance);
-	float getVariance();
+	void setVariance(double newVariance);
+	double getVariance();
 
 	Body* getParent();
 	void setParent(Body* newParent);
 private:
 	Body* parent;
-	float value;
-	float variance;
+	double value;
+	double variance;
 	//The values which lead to the necessity of a covariance matrix in Kalman filtering are dealt with in the actual positional values.
 };
